bi_agent/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
openai>=1.30.0
python-dotenv>=1.0.1
pydantic>=2.7.1
openpyxl>=3.1.2
python-multipart>=0.0.9
jinja2>=3.1.4
aiofiles>=23.2.1